+++ src/attributecode/cmd.py
@@ -41,9 +41,13 @@ from attributecode.model import copy_redist_src
 from attributecode.model import get_copy_list
 from attributecode.model import pre_process_and_fetch_license_dict
 from attributecode.model import write_output
-from attributecode.transform import transform_csv_to_csv
-from attributecode.transform import transform_json_to_json
-from attributecode.transform import transform_excel_to_excel
+from attributecode.transform import transform_data
+from attributecode.transform import transform_csv
+from attributecode.transform import transform_json
+from attributecode.transform import transform_excel
+from attributecode.transform import write_csv
+from attributecode.transform import write_json
+from attributecode.transform import write_excel
 from attributecode.transform import Transformer
 from attributecode.util import extract_zip
 from attributecode.util import filter_errors
@@ -771,8 +775,7 @@ def print_config_help(ctx, param, value):
 def transform(location, output, configuration, quiet, verbose):  # NOQA
     """
 Transform the CSV/JSON/XLSX file at LOCATION by applying renamings, filters and checks
-and then write a new CSV/JSON/XLSX to OUTPUT (Format for input and output need to be
-the same).
+and then write a new CSV/JSON/XLSX to OUTPUT.
 
 LOCATION: Path to a CSV/JSON/XLSX file.
 
@@ -783,16 +786,32 @@ OUTPUT: Path to CSV/JSON/XLSX inventory file to create.
     else:
         transformer = Transformer.from_file(configuration)
 
-    if location.endswith('.csv') and output.endswith('.csv'):
-        errors = transform_csv_to_csv(location, output, transformer)
-    elif location.endswith('.json') and output.endswith('.json'):
-        errors = transform_json_to_json(location, output, transformer)
-    elif location.endswith('.xlsx') and output.endswith('.xlsx'):
-        errors = transform_excel_to_excel(location, output, transformer)
-    else:
-        msg = 'Extension for the input and output need to be the same.'
+    if not transformer:
+        msg = 'Cannot transform without Transformer'
         click.echo(msg)
-        sys.exit()
+        sys.exit(1)
+
+    errors = []
+    updated_data = []
+    new_data = []
+
+    if location.endswith('.csv'):
+        new_data, errors = transform_csv(location)
+    elif location.endswith('.json'):
+        errors = transform_json(location)
+    elif location.endswith('.xlsx'):
+        errors = transform_excel(location)
+
+    if not errors:
+        updated_data, errors = transform_data(new_data, transformer)
+
+    if not errors:
+        if output.endswith('.csv'):
+            write_csv(output, updated_data)
+        elif output.endswith('.json'):
+            write_json(output, updated_data)
+        else:
+            write_excel(output, updated_data)
 
     if not quiet:
         print_version()