+++ tests/test_cmd.py
@@ -136,8 +136,6 @@ def test_report_errors_with_verbose_flag(capsys):
         'DEBUG: msg4',
         'NOTSET: msg4'
     ]
-    print("@@@@@@@@@@@@@@@@@@@@@@@@")
-    print(out.splitlines(False))
     assert expected_out == out.splitlines(False)
     assert '' == err
 
@@ -334,10 +332,6 @@ def check_about_stdout(options, expected_loc, regen=False):
     with open(expected_file, 'r') as ef:
         expected = ef.read()
 
-    print("!!!!!!!!!!!!!!!!!!!!")
-    print(expected.splitlines(False))
-    print("#####################")
-    print(result.output.splitlines(False))
     assert expected.splitlines(False) == result.output.splitlines(False)